Fill database name in sysview tests
