+++ ydb/core/sys_view/ut/ya.make
@@ -30,6 +30,7 @@ SRCS(
     ut_counters.cpp
     ut_labeled.cpp
     ut_registry.cpp
+    ut_show_create.cpp
 )
 
 END()
